feat: setup complete set request system
