+++ src/app/(main)/(buzzer)/buzzer-box/page.tsx
@@ -1,8 +1,7 @@
 import BoxPage from "@/components/pages/buzzer/Box";
-// import { prismaClient } from "@/src/utils/clients";
+import { prismaClient } from "@/src/utils/clients";
 
 export default async function Page() {
-  // const sets = await prismaClient.set.findMany({});
-  // console.log(sets.length);
-  return <BoxPage sets={[]} />;
+  const sets = await prismaClient.set.findManyCompleteSets();
+  return <BoxPage sets={sets} />;
 }
